final version of week 5

+++ app/Http/Controllers/ReceipeController.php
@@ -2,11 +2,18 @@
 
 namespace App\Http\Controllers;
 
+use App\Category;
 use App\Receipe;
 use Illuminate\Http\Request;
 
 class ReceipeController extends Controller
 {
+     public function __construct()
+    {
+        // only
+        // except
+        $this->middleware('auth')->except('create');
+    }
     /**
      * Display a listing of the resource.
      *
@@ -14,7 +21,9 @@ class ReceipeController extends Controller
      */
     public function index()
     {
-       $data = Receipe::all();
+       // $data = Receipe::all();
+        $data = Receipe::where('author_id', auth()->id())->get();
+       // dd(auth()->user());
        return view('home', compact('data'));
     }
 
@@ -25,7 +34,9 @@ class ReceipeController extends Controller
      */
     public function create()
     {
-        return view('create');
+
+        $category = Category::all();
+        return view('create',compact('category'));
     }
 
     /**
@@ -36,11 +47,13 @@ class ReceipeController extends Controller
      */
     public function store()
     {
+        // dd(request()->all());
 
         $validatedData = request()->validate([
             'name'        => 'required',
             'ingredients' => 'required',
             'category'    => 'required',
+            // 'author_id'    => 'required',
         ]);
 
         // $receipe = new Receipe();
@@ -57,9 +70,8 @@ class ReceipeController extends Controller
         //     'category'    => request()->category,
         // ]);
 
-        Receipe::create($validatedData);
+        Receipe::create($validatedData + ['author_id' => auth()->id()]);
         // $receipe->update($validateData);
-
         return redirect('receipe');
     }
 
@@ -71,6 +83,12 @@ class ReceipeController extends Controller
      */
     public function show(Receipe $receipe)
     {
+        // if ($receipe->author_id != auth()->id()){
+        //    abort(404); 
+        // }
+
+        // dd($receipe->categories->name);
+        $this->authorize('view',$receipe);
         return view('show',compact('receipe'));
     }
 
@@ -82,7 +100,9 @@ class ReceipeController extends Controller
      */
     public function edit(Receipe $receipe)
     {
-        return view('edit',compact('receipe'));
+        $this->authorize('view',$receipe);
+        $category = Category::all();
+        return view('edit',compact('receipe','category'));
     }
 
     /**
@@ -94,6 +114,7 @@ class ReceipeController extends Controller
      */
     public function update(Receipe $receipe)
     {
+        $this->authorize('view',$receipe);
         $validatedData = request()->validate([
             'name'        => 'required',
             'ingredients' => 'required',
@@ -119,6 +140,7 @@ class ReceipeController extends Controller
      */
     public function destroy(Receipe $receipe)
     {
+        $this->authorize('view',$receipe);
        $receipe->delete();
        return redirect('receipe'); 
     }
